remove c++11 features
