+++ ext/v8/rr.h
@@ -30,7 +30,6 @@ inline VALUE not_implemented(const char* message) {
 #include "handles.h"
 #include "context.h"
 
-#include "uint32.h"
 #include "value.h"
 #include "backref.h"
 
@@ -38,6 +37,11 @@ inline VALUE not_implemented(const char* message) {
 #include "array.h"
 #include "primitive.h"
 #include "number.h"
+#include "integer.h"
+#include "int32.h"
+#include "uint32.h"
+
+
 #include "external.h"
 // This one is named v8_string to avoid name collisions with C's string.h
 #include "name.h"
